% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Bed_Unit_Utilization.R
\name{Bed_Unit_Utilization}
\alias{Bed_Unit_Utilization}
\title{Bed Utilizations}
\usage{
Bed_Unit_Utilization(
  clarity_api = get_clarity_api(e = rlang::caller_env()),
  app_env = get_app_env(e = rlang::caller_env())
)
}
\description{
Bed Utilizations
}
