% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BoSGoals}
\alias{BoSGoals}
\title{Goal metrics for the Balance of State}
\format{
A tibble with 14 rows and 11 variables:
\describe{
\item{SummaryMeasure}{ The KPM description}
\item{Measure}{ The KPM short name}
\item{Operator}{ The Operator pertaining to successful performance for this metric}
\item{1,2,8,4,3,13,12,9}{ Thresholds for the specific Project Types}
}
}
\usage{
BoSGoals
}
\description{
Thresholds for Balance of State key performance metrics pertaining to specific project types (as column headers)
}
\keyword{datasets}
