% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.2_covid19_utils.R
\name{c19_plot}
\alias{c19_plot}
\title{Create covid19 Prioritization plots}
\usage{
c19_plot(x, y_col, date_breaks = "2 months", palette = "GnBu")
}
\arguments{
\item{x}{\code{(data.frame)} priority or covid19_status}

\item{y_col}{\code{(symbol)} Column that will populate the Y axis values}

\item{date_breaks}{A string giving the distance between breaks like "2
weeks", or "10 years". If both \code{breaks} and \code{date_breaks} are
specified, \code{date_breaks} wins. Valid specifications are 'sec', 'min',
'hour', 'day', 'week', 'month' or 'year', optionally followed by 's'.}

\item{palette}{If a string, will use that named palette. If a number, will index into
the list of palettes of appropriate \code{type}. The list of available palettes can found
in the Palettes section.}
}
\value{
\code{(Plot)}
}
\description{
Create covid19 Prioritization plots
}
