% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_imports.R
\name{clarity_url}
\alias{clarity_url}
\title{Retrieve the Clarity URL from options.
See \code{?setup_RmData} for details.}
\usage{
clarity_url()
}
\value{
\code{(character)}
}
\description{
Retrieve the Clarity URL from options.
See \code{?setup_RmData} for details.
}
