% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Client_add_UniqueID}
\alias{Client_add_UniqueID}
\title{Add the UniqueID to the Client export}
\usage{
Client_add_UniqueID(
  Client,
  Client_extras,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{Client}{\code{(data.frame)} Client HUD CSV Export item}

\item{Client_extras}{\code{(data.frame)} A custom look linking PersonalID & UniqueID}

\item{app_env}{}
}
\value{

}
\description{
Add the UniqueID to the Client export
}
