% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{copy_lgl}
\alias{copy_lgl}
\title{Copy files to directory}
\usage{
copy_lgl(files, dir, overwrite)
}
\arguments{
\item{files}{\code{(character)} of file paths}

\item{dir}{\code{(character)} path to directory where files are to be copied}

\item{overwrite}{logical; should existing destination files be overwritten?}
}
\description{
Copy a character vector of file paths to a directory with informative messages
}
