% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{create_APs}
\alias{create_APs}
\title{create_APs}
\usage{
create_APs(provider_extras, dirs)
}
\arguments{
\item{dirs}{}
}
\value{
\code{(data.frame)}
}
\description{
Create data.frame of Coordinated Entry Access Points with info about the Counties & Populations Served
}
