% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_plots.R
\name{data_quality_plots}
\alias{data_quality_plots}
\title{Render Data Quality plots for RminorElevated}
\usage{
data_quality_plots(
  served_in_date_range,
  Referrals,
  Project,
  dq_past_year,
  dq_unsheltered,
  rm_dates,
  data_APs,
  ProjectDisplay,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{app_env}{}
}
\value{

}
\description{
Render Data Quality plots for RminorElevated
}
