% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_summary.R
\name{data_quality_summary}
\alias{data_quality_summary}
\title{Render Data Quality plots & tables for RminorElevated dq_system_wide}
\usage{
data_quality_summary(
  served_in_date_range,
  Referrals,
  Project,
  dq_past_year,
  dq_eligibility_detail,
  data_APs,
  rm_dates,
  vars,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{app_env}{}
}
\value{

}
\description{
Render Data Quality plots & tables for RminorElevated dq_system_wide
}
