% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update.R
\name{data_ready}
\alias{data_ready}
\title{Is the data in \code{dir} updated?}
\usage{
data_ready(dir = clarity.looker::dirs$export)
}
\arguments{
\item{dir}{\code{(character)} path to directory}
}
\value{
\code{(logical)}
}
\description{
Is the data in \code{dir} updated?
}
