% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_check_eligibility}
\alias{dq_check_eligibility}
\title{Check Eligibility}
\usage{
dq_check_eligibility(
  served_in_date_range,
  mahoning_projects,
  vars,
  rm_dates,
  app_env = get_app_env(e = rlang::caller_env()),
  detail = FALSE
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{mahoning_projects}{\code{(named numeric)} Vector of ProjectIDs associated with projects operating in Mahoning. Created in \code{load_export}.}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}

\item{detail}{\code{(logical)} Whether to return eligibility detail}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
The Residence Prior may suggest that the project is serving ineligible households, the household was entered into the wrong project, or the Residence Prior at Entry is incorrect.
}
\seealso{
Other DQ: Check Eligibility: 
\code{\link{dq_dkr_destination}()},
\code{\link{dq_missing_destination}()},
\code{\link{dq_rent_paid_no_move_in}()}
}
\concept{DQ: Check Eligibility}
