% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_dkr_client_veteran_info}
\alias{dq_dkr_client_veteran_info}
\title{Check for Dont Know/Refused Wars/Branch/Discharge on clients who are Veterans}
\usage{
dq_dkr_client_veteran_info(
  ssvf_served_in_date_range,
  vars,
  guidance,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{ssvf_served_in_date_range}{See \code{ssvf_served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Check for Dont Know/Refused Wars/Branch/Discharge on clients who are Veterans
}
\seealso{
Other DQ: SSVF Checks: 
\code{\link{dq_ssvf_missing_percent_ami}()},
\code{\link{dq_veteran_missing_branch}()},
\code{\link{dq_veteran_missing_discharge_status}()},
\code{\link{dq_veteran_missing_year_entered}()},
\code{\link{dq_veteran_missing_year_separated}()}

Other DQ: EE Checks: 
\code{\link{dq_conflicting_income}()},
\code{\link{dq_duplicate_ees}()},
\code{\link{dq_future_ees}()},
\code{\link{dq_future_exits}()},
\code{\link{dq_missing_income}()},
\code{\link{dq_ssvf_missing_percent_ami}()},
\code{\link{dq_veteran_missing_branch}()},
\code{\link{dq_veteran_missing_discharge_status}()},
\code{\link{dq_veteran_missing_year_entered}()},
\code{\link{dq_veteran_missing_year_separated}()}
}
\concept{DQ: EE Checks}
\concept{DQ: SSVF Checks}
