% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_dkr_living_situation}
\alias{dq_dkr_living_situation}
\title{Find Don't Know/Refused Living Situation}
\usage{
dq_dkr_living_situation(
  served_in_date_range,
  vars,
  rm_dates = NULL,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Don't Know/Refused Living Situation
}
\seealso{
Other DQ: Missing Data at Entry: 
\code{\link{dq_detail_missing_disabilities}()},
\code{\link{dq_dkr_LoS}()},
\code{\link{dq_dkr_months_times_homeless}()},
\code{\link{dq_dkr_residence_prior}()},
\code{\link{dq_invalid_months_times_homeless}()},
\code{\link{dq_missing_LoS}()},
\code{\link{dq_missing_approx_date_homeless}()},
\code{\link{dq_missing_living_situation}()},
\code{\link{dq_missing_months_times_homeless}()},
\code{\link{dq_missing_previous_street_ESSH}()},
\code{\link{dq_missing_residence_prior}()}
}
\concept{DQ: Missing Data at Entry}
