% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_dkr_months_times_homeless}
\alias{dq_dkr_months_times_homeless}
\title{Find Don't Know/Refused Months or Times Homeless}
\usage{
dq_dkr_months_times_homeless(
  served_in_date_range,
  vars,
  rm_dates = NULL,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Don't Know/Refused Months or Times Homeless
}
\seealso{
Other Clarity Checks: 
\code{\link{dq_aps_with_ees}()},
\code{\link{dq_check_disability_ssi}()},
\code{\link{dq_check_eligibility}()},
\code{\link{dq_conflicting_hi_ee}()},
\code{\link{dq_conflicting_income}()},
\code{\link{dq_conflicting_unlikely_ncbs}()},
\code{\link{dq_detail_missing_disabilities}()},
\code{\link{dq_dkr_LoS}()},
\code{\link{dq_dkr_destination}()},
\code{\link{dq_dkr_living_situation}()},
\code{\link{dq_dkr_residence_prior}()},
\code{\link{dq_dob}()},
\code{\link{dq_dose_date_error}()},
\code{\link{dq_duplicate_ees}()},
\code{\link{dq_ethnicity}()},
\code{\link{dq_future_ees}()},
\code{\link{dq_future_exits}()},
\code{\link{dq_gender}()},
\code{\link{dq_hh_children_only}()},
\code{\link{dq_hh_missing_rel_to_hoh}()},
\code{\link{dq_hh_no_oh}()},
\code{\link{dq_hh_too_many_hohs}()},
\code{\link{dq_incorrect_path_contact_date}()},
\code{\link{dq_invalid_months_times_homeless}()},
\code{\link{dq_mahoning_ce_60_days}()},
\code{\link{dq_missing_LoS}()},
\code{\link{dq_missing_approx_date_homeless}()},
\code{\link{dq_missing_client_location}()},
\code{\link{dq_missing_county_prior}()},
\code{\link{dq_missing_county_served}()},
\code{\link{dq_missing_destination}()},
\code{\link{dq_missing_hi_entry}()},
\code{\link{dq_missing_hi_exit}()},
\code{\link{dq_missing_income}()},
\code{\link{dq_missing_living_situation}()},
\code{\link{dq_missing_ncbs}()},
\code{\link{dq_missing_path_contact}()},
\code{\link{dq_missing_previous_street_ESSH}()},
\code{\link{dq_missing_residence_prior}()},
\code{\link{dq_missing_vaccine_current}()},
\code{\link{dq_missing_vaccine_exited}()},
\code{\link{dq_name}()},
\code{\link{dq_overlaps_psh}()},
\code{\link{dq_overlaps_rrh}()},
\code{\link{dq_overlaps_same_day}()},
\code{\link{dq_overlaps}()},
\code{\link{dq_path_SOAR_missing_at_exit}()},
\code{\link{dq_path_enrolled_missing}()},
\code{\link{dq_path_no_status_at_exit}()},
\code{\link{dq_path_reason_missing}()},
\code{\link{dq_path_status_determination}()},
\code{\link{dq_ph_without_spdats}()},
\code{\link{dq_project_small}()},
\code{\link{dq_psh_check_exit_destination}()},
\code{\link{dq_psh_incorrect_destination}()},
\code{\link{dq_psh_missing_project_stay}()},
\code{\link{dq_race}()},
\code{\link{dq_referrals_on_hh_members_ssvf}()},
\code{\link{dq_rent_paid_no_move_in}()},
\code{\link{dq_rrh_check_exit_destination}()},
\code{\link{dq_rrh_missing_project_stay}()},
\code{\link{dq_services_on_hh_members_ssvf}()},
\code{\link{dq_services_on_non_hoh}()},
\code{\link{dq_sh_check_exit_destination}()},
\code{\link{dq_sh_missing_project_stay}()},
\code{\link{dq_ssn}()},
\code{\link{dq_th_check_exit_destination}()},
\code{\link{dq_th_missing_project_stay}()},
\code{\link{dq_th_stayers_bos}()},
\code{\link{dq_veteran}()}

Other DQ: Missing Data at Entry: 
\code{\link{dq_detail_missing_disabilities}()},
\code{\link{dq_dkr_LoS}()},
\code{\link{dq_dkr_living_situation}()},
\code{\link{dq_dkr_residence_prior}()},
\code{\link{dq_invalid_months_times_homeless}()},
\code{\link{dq_missing_LoS}()},
\code{\link{dq_missing_approx_date_homeless}()},
\code{\link{dq_missing_living_situation}()},
\code{\link{dq_missing_months_times_homeless}()},
\code{\link{dq_missing_previous_street_ESSH}()},
\code{\link{dq_missing_residence_prior}()}
}
\concept{Clarity Checks}
\concept{DQ: Missing Data at Entry}
