% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_incorrect_path_contact_date}
\alias{dq_incorrect_path_contact_date}
\title{PATH: Incorrect PATH Contact Date}
\usage{
dq_incorrect_path_contact_date(
  served_in_date_range,
  Contacts,
  rm_dates,
  vars,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{Contacts}{\code{(data.frame)} From the HUD CSV Export}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Every adult or head of household should have a Living Situation contact record where the Contact Date matches the Entry Date. This would represent the initial contact made with the client.
}
\details{
client is adult/hoh, has a contact record, and the first record in the EE does not equal the Entry Date ->  error
}
\seealso{
Other DQ: Path Checks: 
\code{\link{dq_missing_path_contact}()},
\code{\link{dq_path_SOAR_missing_at_exit}()},
\code{\link{dq_path_enrolled_missing}()},
\code{\link{dq_path_missing_los_res_prior}()},
\code{\link{dq_path_no_status_at_exit}()},
\code{\link{dq_path_reason_missing}()},
\code{\link{dq_path_status_determination}()},
\code{\link{dq_project_small}()}
}
\concept{DQ: Path Checks}
