% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_missing_destination}
\alias{dq_missing_destination}
\title{Missing Destination}
\usage{
dq_missing_destination(
  served_in_date_range,
  mahoning_projects,
  vars,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{mahoning_projects}{\code{(named numeric)} Vector of ProjectIDs associated with projects operating in Mahoning. Created in \code{load_export}.}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
It is widely understood that not every client will complete an exit interview, especially for high-volume emergency shelters. A few warnings for Missing Destination is no cause for concern, but if there is a large number this will surface these errors.
}
\seealso{
Other DQ: Check Eligibility: 
\code{\link{dq_check_eligibility}()},
\code{\link{dq_dkr_destination}()},
\code{\link{dq_rent_paid_no_move_in}()}
}
\concept{DQ: Check Eligibility}
