% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_missing_income}
\alias{dq_missing_income}
\title{Find Conflicting Income yes/no at Entry or Exit}
\usage{
dq_missing_income(
  served_in_date_range,
  IncomeBenefits,
  vars,
  guidance,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{IncomeBenefits}{\code{(data.frame)} From the HUD CSV Export}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Please enter the data for this item by clicking into the Entry or Exit pencil on the given Client ID on the appropriate program stay.
}
\seealso{
Other DQ: Income Checks: 
\code{\link{dq_conflicting_income}()}

Other DQ: EE Checks: 
\code{\link{dq_conflicting_income}()},
\code{\link{dq_duplicate_ees}()},
\code{\link{dq_future_ees}()},
\code{\link{dq_future_exits}()}
}
\concept{DQ: EE Checks}
\concept{DQ: Income Checks}
