% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_missing_months_times_homeless}
\alias{dq_missing_months_times_homeless}
\title{Find Missing Months or Times Homeless}
\usage{
dq_missing_months_times_homeless(
  served_in_date_range,
  vars,
  guidance = NULL,
  rm_dates = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Missing Months or Times Homeless
}
\seealso{
Other DQ: Missing Data at Entry: 
\code{\link{dq_date_homeless_after_entry}()},
\code{\link{dq_detail_missing_disabilities}()},
\code{\link{dq_dkr_LoS}()},
\code{\link{dq_dkr_months_times_homeless}()},
\code{\link{dq_dkr_prior_living_situation}()},
\code{\link{dq_invalid_months_times_homeless}()},
\code{\link{dq_missing_approx_date_homeless}()},
\code{\link{dq_missing_living_situation}()},
\code{\link{dq_missing_previous_street_ESSH}()},
\code{\link{dq_missing_prior_living_situation}()},
\code{\link{dq_months_homeless_tbd}()},
\code{\link{dq_sp_missing_LoS}()}
}
\concept{DQ: Missing Data at Entry}
