% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_missing_ncbs}
\alias{dq_missing_ncbs}
\title{Find Missing Non-Cash Benefits (NCBS) at Entry/Exit}
\usage{
dq_missing_ncbs(
  served_in_date_range,
  IncomeBenefits,
  vars,
  guidance,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{IncomeBenefits}{\code{(data.frame)} From the HUD CSV Export}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
This data element is required to be collected at project Entry or Exit. Please click into the client's Entry/Exit pencil to save this data to HMIS.
}
\seealso{
Other DQ: Non-Cash Benefit Checks: 
\code{\link{dq_conflicting_unlikely_ncbs}()}
}
\concept{DQ: Non-Cash Benefit Checks}
