% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_missing_vaccine_current}
\alias{dq_missing_vaccine_current}
\title{Find Missing vaccine data for enrolled clients}
\usage{
dq_missing_vaccine_current(
  served_in_date_range,
  vars,
  dose_counts,
  doses = NULL,
  mahoning_projects = NULL,
  hc = NULL,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{doses}{\code{(data.frame)} See the \verb{[["HUD Extras"]]$Client_Doses_extra} method in the instantiated clarity_api object}

\item{mahoning_projects}{\code{(numeric)} from \code{load_export}}

\item{hc}{\code{(list)} from \code{dates}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Missing vaccine data for enrolled clients
}
