% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_path_missing_los_res_prior}
\alias{dq_path_missing_los_res_prior}
\title{PATH: Missing Residence Prior Length of Stay}
\usage{
dq_path_missing_los_res_prior(
  served_in_date_range,
  vars,
  guidance,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
PATH: Missing Residence Prior Length of Stay
}
\seealso{
Other DQ: Path Checks: 
\code{\link{dq_incorrect_path_contact_date}()},
\code{\link{dq_missing_path_contact}()},
\code{\link{dq_path_SOAR_missing_at_exit}()},
\code{\link{dq_path_enrolled_missing}()},
\code{\link{dq_path_no_status_at_exit}()},
\code{\link{dq_path_reason_missing}()},
\code{\link{dq_path_status_determination}()},
\code{\link{dqu_project_small}()}
}
\concept{DQ: Path Checks}
