% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_path_SOAR_missing_at_exit}
\alias{dq_path_SOAR_missing_at_exit}
\title{PATH: Missing Connection with SOAR at Exit}
\usage{
dq_path_SOAR_missing_at_exit(
  served_in_date_range,
  Project,
  IncomeBenefits,
  guidance,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{Project}{\code{(data.frame)} of Project with additional features added in \code{load_export}}

\item{IncomeBenefits}{\code{(data.frame)} From the HUD CSV Export}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
PATH: Missing Connection with SOAR at Exit
}
\details{
adult & Connection w/ SOAR is null or DNC -> error
}
\seealso{
Other DQ: Path Checks: 
\code{\link{dq_incorrect_path_contact_date}()},
\code{\link{dq_missing_path_contact}()},
\code{\link{dq_path_enrolled_missing}()},
\code{\link{dq_path_missing_los_res_prior}()},
\code{\link{dq_path_no_status_at_exit}()},
\code{\link{dq_path_reason_missing}()},
\code{\link{dq_path_status_determination}()},
\code{\link{dq_project_small}()}
}
\concept{DQ: Path Checks}
