% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_project_small}
\alias{dq_project_small}
\title{Return a subset of Project data}
\usage{
dq_project_small(Projec)
}
\arguments{
\item{Project}{\code{(data.frame)} of Project with additional features added in \code{load_export}}
}
\value{
\code{(data.frame)} with ProjectID, ProjectName, ProjectCounty
}
\description{
Return a subset of Project data
}
\seealso{
Other DQ: Path Checks: 
\code{\link{dq_incorrect_path_contact_date}()},
\code{\link{dq_missing_path_contact}()},
\code{\link{dq_path_SOAR_missing_at_exit}()},
\code{\link{dq_path_enrolled_missing}()},
\code{\link{dq_path_missing_los_res_prior}()},
\code{\link{dq_path_no_status_at_exit}()},
\code{\link{dq_path_reason_missing}()},
\code{\link{dq_path_status_determination}()}
}
\concept{DQ: Path Checks}
