% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils_ServicePoint.R
\name{dq_sp_hh_missing_rel_to_hoh}
\alias{dq_sp_hh_missing_rel_to_hoh}
\title{Find Households with Missing Relationship to Head of Household}
\usage{
dq_sp_hh_missing_rel_to_hoh(
  served_in_date_range,
  vars,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Households with Missing Relationship to Head of Household
}
\seealso{
Other ServicePoint Checks: 
\code{\link{dq_overlaps}()},
\code{\link{dq_sp_hh_too_many_hohs}()},
\code{\link{dq_sp_missing_LoS}()},
\code{\link{dq_sp_missing_client_location}()},
\code{\link{overlaps_rrh}()}

Other DQ: Household Checks: 
\code{\link{dq_hh_children_only}()},
\code{\link{dq_hh_no_hoh}()},
\code{\link{dq_referrals_on_hh_members_ssvf}()},
\code{\link{dq_services_on_hh_members_ssvf}()},
\code{\link{dq_services_on_non_hoh}()},
\code{\link{dq_sp_hh_too_many_hohs}()}
}
\concept{DQ: Household Checks}
\concept{ServicePoint Checks}
