% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_sp_incorrect_ee_type}
\alias{dq_sp_incorrect_ee_type}
\title{Incorrect Entry/Exit Types}
\usage{
dq_sp_incorrect_ee_type(server_in_date_range)
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Incorrect Entry/Exit Types
}
\concept{data_quality_servicepoint}
