% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils_ServicePoint.R
\name{dq_sp_missing_client_location}
\alias{dq_sp_missing_client_location}
\title{Find missing client locations}
\usage{
dq_sp_missing_client_location(
  served_in_date_range,
  vars,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find missing client locations
}
\seealso{
Other ServicePoint Checks: 
\code{\link{dq_overlaps}()},
\code{\link{dq_sp_hh_missing_rel_to_hoh}()},
\code{\link{dq_sp_hh_too_many_hohs}()},
\code{\link{dq_sp_missing_LoS}()},
\code{\link{overlaps}()}
}
\concept{ServicePoint Checks}
