% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils_ServicePoint.R
\name{dq_sp_missing_LoS}
\alias{dq_sp_missing_LoS}
\title{Find Missing Length of Stay}
\usage{
dq_sp_missing_LoS(
  served_in_date_range,
  vars,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Missing Length of Stay
}
\seealso{
Other ServicePoint Checks: 
\code{\link{dq_overlaps}()},
\code{\link{dq_sp_hh_missing_rel_to_hoh}()},
\code{\link{dq_sp_hh_too_many_hohs}()},
\code{\link{dq_sp_missing_client_location}()},
\code{\link{overlaps}()}

Other DQ: Missing Data at Entry: 
\code{\link{dq_detail_missing_disabilities}()},
\code{\link{dq_dkr_LoS}()},
\code{\link{dq_dkr_months_times_homeless}()},
\code{\link{dq_dkr_prior_living_situation}()},
\code{\link{dq_invalid_months_times_homeless}()},
\code{\link{dq_missing_approx_date_homeless}()},
\code{\link{dq_missing_living_situation}()},
\code{\link{dq_missing_months_times_homeless}()},
\code{\link{dq_missing_previous_street_ESSH}()},
\code{\link{dq_missing_prior_living_situation}()}
}
\concept{DQ: Missing Data at Entry}
\concept{ServicePoint Checks}
