% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_ssn}
\alias{dq_ssn}
\title{Data quality report on SSN Validity}
\usage{
dq_ssn(served_in_date_range, guidance, vars)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{guidance}{\code{(list)}}

\item{vars}{\code{(list)}}
}
\value{
\code{(data.frame)} with \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Data quality report on SSN Validity
}
