% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_th_check_exit_destination}
\alias{dq_th_check_exit_destination}
\title{Find Incorrect Exits in TH}
\usage{
dq_th_check_exit_destination(
  served_in_date_range,
  vars,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Incorrect Exits in TH
}
\seealso{
Other data_quality_destinations: 
\code{\link{dq_psh_check_exit_destination}()},
\code{\link{dq_psh_incorrect_destination}()},
\code{\link{dq_psh_missing_project_stay}()},
\code{\link{dq_rrh_check_exit_destination}()},
\code{\link{dq_rrh_missing_project_stay}()},
\code{\link{dq_sh_check_exit_destination}()},
\code{\link{dq_sh_missing_project_stay}()},
\code{\link{dq_th_missing_project_stay}()}
}
\concept{data_quality_destinations}
