% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{enrolled_in}
\alias{enrolled_in}
\title{Filter for Enrollments in a Specific Project Type}
\usage{
enrolled_in(
  served_in_date_range,
  type = c(ES = 1, TH = 2, PSH = 3, SO = 4, ServicesOnly = 6, Other = 7, SH = 8, PHHO =
    9, PHHS = 10, DS = 11, HP = 12, RRH = 13, CE = 14)[13],
  has_movein = FALSE
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{type}{\code{(numeric)} ProjectType. For full project type names see \verb{hud.extract::hud_translations$[["2.02.6 ProjectType"]](table = TRUE)}}
}
\value{
\code{(data.frame)} with \code{PersonalID} for all Enrollees in the ProjectType, their \code{MoveInDateAdjust}, the \code{TimeInterval} for which they were in the Project, and the \code{ProjectName}
}
\description{
Filter for Enrollments in a Specific Project Type
}
