% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Enrollment_add_AgeAtEntry}
\alias{Enrollment_add_AgeAtEntry}
\title{Add AgeAtEntry to Enrollment}
\usage{
Enrollment_add_AgeAtEntry(Enrollment, Client)
}
\arguments{
\item{Client}{}
}
\value{
\code{(data.frame)} Enrollment with \code{AgeAtEntry} column
}
\description{
AgeAtEntry is the time elapsed from the Date of Birth \code{DOB} to the \code{EntryDate}
}
