% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{Enrollment_add_ClientLocation}
\alias{Enrollment_add_ClientLocation}
\title{Add Client Location to Enrollment}
\usage{
Enrollment_add_ClientLocation(Enrollment, EnrollmentCoC)
}
\arguments{
\item{Enrollment}{\code{(data.frame)} From the HUD CSV Export}

\item{EnrollmentCoC}{\code{(data.frame)} From the HUD CSV Export}
}
\value{
\code{(data.frame)} Enrollment with \code{ClientLocation} column
}
\description{
Add Client Location to Enrollment
}
