% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Enrollment_add_Exit}
\alias{Enrollment_add_Exit}
\title{Add Exit data to Enrollments}
\usage{
Enrollment_add_Exit(Enrollment, Exit)
}
\arguments{
\item{Exit}{}
}
\value{
\code{(data.frame)} Enrollments with \code{ExitDate}, \code{Destination}, \code{OtherDestination} and derived column \code{ExitAdjust}
}
\description{
Add Exit data to Enrollments
}
