% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Enrollment_add_Household}
\alias{Enrollment_add_Household}
\title{Add Household Information to Enrollment from Project}
\usage{
Enrollment_add_Household(Enrollment, Project, hc)
}
\arguments{
\item{Enrollment}{with Exit data. See \code{Enrollment_add_Exit}}

\item{hc}{From \code{dates}}
}
\value{
\code{(data.frame)} of Enrollment with Household Columns \code{MoveInDateAdjust} appended
}
\description{
Add Household Information to Enrollment from Project
}
