% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{Enrollment_add_HousingStatus}
\alias{Enrollment_add_HousingStatus}
\title{Add HousingStatus indicating the client's current housing status and Situation with details on that status.}
\usage{
Enrollment_add_HousingStatus(
  Enrollment_extra_Client_Exit_HH_CL_AaE,
  PH = data_types$Project$ProjectType$ph,
  Referrals,
  prioritization_colors,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{PH}{ProjectType codes considered Permanently Housed.  See \code{HMIS::hud_translations$ProjectType(table = TRUE)} & \code{?data_types}}

\item{Referrals}{\code{(data.frame)} From the Clarity Looker API Extras}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
Input data with HousingStatus & Situation column
}
\description{
Add HousingStatus indicating the client's current housing status and Situation with details on that status.
}
