% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{EnrollmentCoC_RemoveCoCCodes}
\alias{EnrollmentCoC_RemoveCoCCodes}
\title{Remove specific CoCCode's from EnrollmentCoC}
\usage{
EnrollmentCoC_RemoveCoCCodes(EnrollmentCoC, codes_to_remove = c("Default"))
}
\arguments{
\item{EnrollmentCoC}{\code{(data.frame)} HUD CSV Item}

\item{codes_to_remove}{\code{(character)} codes to remove by filtering the \code{CoCCode} column}
}
\value{
\code{(data.frame)} without the entries for the specified \code{CoCCode}s
}
\description{
Remove specific CoCCode's from EnrollmentCoC
}
