% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{filter_dupe_last_EnrollmentID}
\alias{filter_dupe_last_EnrollmentID}
\title{Filter for the Last Enrollment ID}
\usage{
filter_dupe_last_EnrollmentID(.data, key, EnrollmentID)
}
\arguments{
\item{.data}{\code{(data.frame)} Data with duplicates}

\item{key}{\code{(name)} of the column key that will be grouped by and for which at least one observation will be preserved.}

\item{EnrollmentID}{\code{(name)} of EnrollmentID column that will be filtered for the maximum (latest)}
}
\value{
\code{(data.frame)} With only the latest Enrollment
}
\description{
The maximum EnrollmentID will always be the latest in Clarity. This function will find the last EnrollmentID per \code{key}
}
