% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{Funder_VA_ProjectID}
\alias{Funder_VA_ProjectID}
\title{Funder_VA}
\usage{
Funder_VA_ProjectID(x, ids = c(27, 30, 33, 37:42, 45))
}
\arguments{
\item{x}{\code{(data.frame)} Funder Export Object}

\item{ids}{\code{(numeric)} All VA Related IDs}
}
\value{
\code{(data.frame)} with Project ID
}
\description{
This filters for VA Funders
}
