% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geocodes}
\alias{geocodes}
\title{Geocodes for the Balance of State}
\format{
A tibble with 123 rows and 5 variables:
\describe{
\item{GeographicCode}{ Geocode}
\item{State}{ State}
\item{Name}{ the City/County Name}
\item{Type}{ the type (City/County)}
\item{County}{ the County (derived column)}
}
}
\source{
\href{https://www.hud.gov/sites/dfiles/CPD/documents/FY-2021-GeoCodes-with-PPRN_Final.pdf}{HUD Geocodes PDF FY 2021}. See \link[hud.extract]{hud_geocodes} for instructions on extracting these in subsequent years.
}
\usage{
geocodes
}
\description{
Maps geocodes to the County/City's of Ohio
}
\keyword{datasets}
