% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{get_app_env}
\alias{get_app_env}
\title{Find the \code{app_env} R6 object}
\usage{
get_app_env(
  nm = "rm_env",
  e = rlang::caller_env(),
  ifnotfound = stop("app_env object not found")
)
}
\arguments{
\item{nm}{The name of the instantiated object}

\item{e}{The environment in which to search}
}
\description{
Find the \code{app_env} R6 object
}
\details{
Searches the \code{RmData} package environment and then the environment specified
}
\seealso{
Other Get R6 Classes: 
\code{\link{get_clarity_api}()}
}
\concept{Get R6 Classes}
