% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{get_clarity_api}
\alias{get_clarity_api}
\title{Find the \code{clarity_api} R6 object}
\usage{
get_clarity_api(
  nm = "cl_api",
  e = rlang::caller_env(),
  ifnotfound = stop("Clarity API object not found")
)
}
\arguments{
\item{nm}{The name of the instantiated object}

\item{e}{The environment in which to search}

\item{ifnotfound}{Warning message if env not found}
}
\value{
\code{clarity_api} R6 object
}
\description{
Find the \code{clarity_api} R6 object
}
\details{
Searches the \code{RmData} package environment and then the environment specified
}
\seealso{
Other Get R6 Classes: 
\code{\link{get_app_env}()}
}
\concept{Get R6 Classes}
