% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export.R
\name{get_export}
\alias{get_export}
\title{Redact PII from Client HUD Export}
\usage{
get_export(hud)
}
\arguments{
\item{Client}{\code{(tibble)} Client HUD Export}
}
\value{
\code{(tibble)} Redacted Client HUD Export
}
\description{
This redacts all PII (except DOB) from Client HUD Export
}
