% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUD_specs}
\alias{HUD_specs}
\title{HUD_specs}
\format{
A tibble with 728 rows and 4 variables:
\describe{
\item{DataElementID}{ The ID number}
\item{DataElement}{ The column name for the variable}
\item{ReferenceNo}{ The value pertaining to the variable}
\item{Description}{ The description of that value}
}
}
\source{
\href{https://hudhdx.info/Resources/Vendors/HMIS_CSV_Specifications_FY2022_v1.2_clean.pdf}{HUD Specs PDF FY 2021}. See \link[hud.extract]{hud_value_tables} for instructions on extracting these in subsequent years.
}
\usage{
HUD_specs
}
\description{
A table of the HUD CSV variable specifications
}
\seealso{
HMIS::hud_translations
}
\keyword{datasets}
