% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{increment}
\alias{increment}
\title{increment steps of 00_copy_images.R}
\usage{
increment(..., cenv = rlang::caller_env())
}
\arguments{
\item{...}{Messages to console}

\item{cenv}{the calling environment}
}
\description{
Provides informative messages and timing for 00_copy_images.R
}
