% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{is_app_env}
\alias{is_app_env}
\title{Is object of class \code{app_env}}
\usage{
is_app_env(x)
}
\arguments{
\item{x}{\code{(any)} The object to check.}
}
\value{
\code{(logical)} Returns \code{TRUE} if \code{x} is of class \code{"app_env"}, otherwise \code{FALSE}.
}
\description{
Check if the provided object is of class \code{app_env}.
}
