% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{is_clarity_api}
\alias{is_clarity_api}
\title{Is object of class \code{clarity_api}}
\usage{
is_clarity_api(x)
}
\arguments{
\item{x}{}
}
\value{
\code{(logical)}
}
\description{
Is object of class \code{clarity_api}
}
