% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{is_sp}
\alias{is_sp}
\title{Is this instance using ServicePoint}
\usage{
is_sp()
}
\value{
\code{(logical)}
}
\description{
Set an option in \code{.Rprofile} using \code{usethis::edit_r_profile('project')} called HMIS which is a list containing two logical values:
\itemize{
\item{\code{Clarity}}{ A logical to indicate whether Clarity is (or has been) used by this CoC}
\item{\code{ServicePoint}}{ A logical to indicate whether Servicepoint is (or has been) used by this CoC}
}
}
