% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{load_enrollment}
\alias{load_enrollment}
\title{Load Enrollment as Enrollment_extra_Client_Exit_HH_CL_AaE}
\usage{
load_enrollment(
  Enrollment,
  EnrollmentCoC,
  Enrollment_extras,
  Exit,
  Client,
  Project,
  Referrals,
  rm_dates,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{Enrollment}{\code{(data.frame)} From the HUD CSV Export}

\item{EnrollmentCoC}{\code{(data.frame)} From the HUD CSV Export}

\item{Enrollment_extras}{From Clarity Looker API Extras}

\item{Exit}{\code{(data.frame)} From the HUD CSV Export}

\item{Client}{\code{(data.frame)} From the HUD CSV Export}

\item{Project}{\code{(data.frame)} From the HUD CSV Export}

\item{Referrals}{\code{(data.frame)} From the Clarity Looker API Extras}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\description{
Load Enrollment as Enrollment_extra_Client_Exit_HH_CL_AaE
}
