% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{load_project}
\alias{load_project}
\title{Load Project & extras pre-requisites for \code{load_export}}
\usage{
load_project(
  Regions,
  ProjectCoC,
  clarity_api = get_clarity_api(e = rlang::caller_env()),
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{Regions}{From public data. See \code{load_public}}
}
\value{
\code{app_env} with data dependencies
}
\description{
Load Project & extras pre-requisites for \code{load_export}
}
