% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{load_referrals}
\alias{load_referrals}
\title{Load Referrals}
\usage{
load_referrals(Referrals, app_env = get_app_env(e = rlang::caller_env()))
}
\arguments{
\item{Referrals}{\code{(data.frame)} From the Clarity Looker API Extras}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Loads Referrals, Referrals_full (unfiltered), and referral_result_summarize with filtering expressions used later
}
