% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{make_linked_df}
\alias{make_linked_df}
\title{Make UniqueID or EnrollmentID into a Clarity hyperlink}
\usage{
make_linked_df(x, ID, unlink = FALSE)
}
\arguments{
\item{x}{\code{(data.frame)} The following columns are required for the specified link type:
\itemize{
\item{\code{PersonalID & UniqueID}}{ for Profile link}
\item{\code{PersonalID & EnrollmentID}}{ for Enrollment link}
}}

\item{ID}{\code{(name)} unquoted of the column to unlink.}

\item{unlink}{\code{(logical)} Whether to turn the link back into the respective columns from which it was made.}
}
\value{
\code{(data.frame)} With \code{UniqueID} or \code{EnrollmentID} as a link
}
\description{
Make UniqueID or EnrollmentID into a Clarity hyperlink
}
