% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{make_profile_link}
\alias{make_profile_link}
\title{Make a Clarity Profile link using the UniqueID and PersonalID}
\usage{
make_profile_link(pid, uid, chr)
}
\arguments{
\item{pid}{\code{(character/data.frame)} Either the \code{PersonalID} column, or the \code{data.frame} with it.}

\item{uid}{\code{(character)} The \code{UniqueID} column, unnecessary to specific if \code{data.frame} supplied to PersonalID}

\item{chr}{\code{(logical)} Whether to output a character or a \code{shiny.tag} if \code{FALSE}. \strong{Default} TRUE}
}
\value{
\code{(character/data.frame/shiny.tag)} If \code{PersonalID} is a character vector (IE nested in a mutate), and \code{chr = TRUE} a character vector, if \code{chr = FALSE} a \code{shiny.tag}. If \code{PersonalID} is a \code{data.frame} a \code{data.frame} with the \code{UniqueID} column replaced with the link to the profile and the \code{UniqueID} as the text
}
\description{
If used in a \link[DT]{datatable}, set \code{escape = FALSE}
}
\examples{
data.frame(a = letters, b = seq_along(letters)) |> dplyr::rowwise() |>  dplyr::mutate(a = make_profile_link(a, b)) |> DT::datatable(escape = FALSE)
}
