% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{make_profile_link_df}
\alias{make_profile_link_df}
\title{Make UniqueID into a Clarity client profile link}
\usage{
make_profile_link_df(x)
}
\arguments{
\item{x}{\code{(data.frame)} must have \code{PersonalID} & \code{UniqueID} columns.}
}
\value{
\code{(data.frame)} With \code{UniqueID} as a profile link and PersonalID removed.
}
\description{
Make UniqueID into a Clarity client profile link
}
\examples{
data.frame(a = letters, b = seq_along(letters)) |> make_profile_link() |>
DT::datatable(escape = FALSE)
}
