% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_imports.R
\name{must_sp}
\alias{must_sp}
\title{This instance must be using ServicePoint, otherwise throw an error.}
\usage{
must_sp(.call = match.call()[[1]])
}
\arguments{
\item{.call}{The function call to display in the error message. Defaults to the current function call.}
}
\value{
\code{(logical)}
}
\description{
Checks if the instance is using ServicePoint. If not, throws an error.
}
