% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update_utils.R
\name{pb_close.character}
\alias{pb_close.character}
\title{Close a Character Progress Bar}
\usage{
\method{pb_close}{character}(pbar, e)
}
\arguments{
\item{pbar}{The progress bar object to close.}

\item{e}{The environment in which to close the progress bar.}
}
\description{
Closes a progress bar for character input.
}
