% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update_utils.R
\name{pb_close.Progress}
\alias{pb_close.Progress}
\title{Close a Progress Object}
\usage{
\method{pb_close}{Progress}(pbar, e)
}
\arguments{
\item{pbar}{The progress bar object to close.}

\item{e}{The environment in which to close the progress bar.}
}
\description{
Closes a progress bar for a \code{Progress} object.
}
