% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update.R
\name{pb_close}
\alias{pb_close}
\title{Close a \link[cli]{cli_progress_bar} or \link[shiny]{Progress} based progress bar}
\usage{
pb_close(pbar)
}
\arguments{
\item{pbar}{\code{(progress bar id (character) or R6 object)}}
}
\description{
Close a \link[cli]{cli_progress_bar} or \link[shiny]{Progress} based progress bar
}
