% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update_utils.R
\name{pb_start}
\alias{pb_start}
\title{Start a Progress Bar}
\usage{
pb_start(
  name = "Daily Update",
  status = NULL,
  type = c("iterator", "tasks", "download", "custom")[1],
  total = NA,
  is_shiny = FALSE,
  e = rlang::caller_env(),
  ...
)
}
\arguments{
\item{name}{\code{(character)} Name of the progress bar.}

\item{status}{\code{(character)} Status message for the progress bar.}

\item{type}{\code{(character)} Type of the progress bar. One of "iterator", "tasks", "download", or "custom".}

\item{total}{\code{(numeric)} \code{max} for \link[shiny]{Progress} & \code{total} for \link[cli]{cli_progress_bar}.}

\item{is_shiny}{\code{(logical)} Whether running in a shiny environment.}

\item{e}{The environment in which to start the progress bar.}

\item{...}{Additional arguments passed to either \link[shiny]{Progress$new} or \link[cli]{cli_progress_bar}.}
}
\value{
An object of class \code{Progress} if \code{is_shiny} is \code{TRUE}, otherwise a \code{cli_progress_bar} object.
}
\description{
Start a \link[cli]{cli_progress_bar} or \link[shiny]{Progress}-based progress bar.
}
