% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Pe_add_GrantType}
\alias{Pe_add_GrantType}
\title{Add GrantType column to provider_extras}
\usage{
Pe_add_GrantType(provider_extras)
}
\arguments{
\item{provider_extras}{}
}
\value{

}
\description{
GrantType indicates if the program is funded by one of HOPWA, PATH, SSVF, or RHY
}
