% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Pe_add_ProjectType}
\alias{Pe_add_ProjectType}
\title{Add ProjectType (dbl) to provider_extras}
\usage{
Pe_add_ProjectType(provider_extras)
}
\value{
\code{(data.frame)}
}
\description{
Add ProjectType (dbl) to provider_extras
}
