% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{pe_create_APs}
\alias{pe_create_APs}
\title{Add Access Points to Provider_extras}
\usage{
pe_create_APs(
  provider_extras,
  ProjectCoC,
  dirs,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{provider_extras}{\code{(data.frame)} provider_extras with Regions, see \code{pe_add_regions}.}

\item{ProjectCoC}{\code{(data.frame)} Data frame containing information about the Project Continuum of Care (CoC).}

\item{dirs}{\code{(list)} List of directories used in the function.}

\item{app_env}{\code{(environment)} The application environment, defaults to the result of \code{get_app_env(e = rlang::caller_env())}.}
}
\value{
\code{(data.frame)} A data frame of Coordinated Entry Access Points with Counties & Populations Served.
}
\description{
Create a data frame of Coordinated Entry Access Points with information about the Counties & Populations Served.
}
