% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{Project_rm_zz}
\alias{Project_rm_zz}
\title{Filter Projects that have been retired (\code{zz}'d)}
\usage{
Project_rm_zz(Project)
}
\arguments{
\item{Project}{\code{(data.frame)} From the HUD CSV Export}
}
\value{

}
\description{
Filter Projects that have been retired (\code{zz}'d)
}
