% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_imports.R
\name{R6classes}
\alias{R6classes}
\title{Roxygen Template for \code{app_env} & \code{clarity_api}}
\arguments{
\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}

\item{clarity_api}{\code{(clarity_api)} R6 Object containing the connection to the Clarity Looker instance.}
}
\description{
Roxygen Template for \code{app_env} & \code{clarity_api}
}
