% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Regions}
\alias{Regions}
\title{Regions}
\format{
A tibble with 82 rows and 3 variables:
\describe{
\item{County}{ The county name}
\item{Region}{ The Region number}
\item{RegionName}{ The Region name}
}
}
\source{
\href{http://hmis.cohhio.org/index.php?pg=kb.page&id=77}{BoS Regions}.
}
\usage{
Regions
}
\description{
A table of the Balance of State Regions
}
\keyword{datasets}
