% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{replace_yes_no}
\alias{replace_yes_no}
\title{replace "yes"/"no" character vector}
\usage{
replace_yes_no(column_name, numeric = TRUE)
}
\arguments{
\item{column_name}{\code{(character)}}

\item{numeric}{\code{(logical)} if \code{numeric = TRUE} numeric, otherwise logical. \strong{Default} \code{TRUE}}
}
\value{
if \code{numeric = TRUE} numeric, otherwise logical.
}
\description{
replace "yes"/"no" character vector
}
