% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoring_rubric}
\alias{scoring_rubric}
\title{The Scoring Rubric}
\format{
A tibble with 85 rows and 6 variables:
\describe{
\item{metric}{ the metric name}
\item{ProjectType}{ The Project Type code to which the metric pertains}
\item{goal_type}{ the direction of the goal}
\item{minimum}{ the minimum cut-off to score}
\item{maximum}{ the maximum value possible}
\item{points}{ the points possible for the metric}
}
}
\usage{
scoring_rubric
}
\description{
Key Performance measures for evaluating projects
}
\keyword{datasets}
