% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ServiceAreas}
\alias{ServiceAreas}
\title{Service Areas}
\format{
A tibble with 88 rows and 13 variables:
\describe{
\item{State}{}
\item{County}{}
\item{SSVFServiceArea}{ the SSVF program for the area}
\item{VAMC}{ the VAMC program for the area}
\item{HomelessPlanningRegion}{ The BoS Region}
\item{SOAR}{ the SOAR program for the region}
\item{Bridges}{ the Bridges region}
\item{BridgesRegCoordinator}{ the Bridges regional coordinator contact details}
\item{CoC}{ the Continuum of Care for the area}
\item{YHDP}{ the YHDP program for the area}
\item{RHY}{ the RHY program for the area}
\item{Unknown}{}
\item{HNHF}{ the HNHF program for the area}
}
}
\usage{
ServiceAreas
}
\description{
Details about each Service area in the Balance of State
}
\keyword{datasets}
