% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{setup_Rm_data}
\alias{setup_Rm_data}
\title{Setup Rm_data options}
\usage{
setup_Rm_data(Clarity = TRUE, ServicePoint = FALSE)
}
\arguments{
\item{Clarity}{\code{(logical)} Does the CoC use Clarity HMIS? \strong{Default} \code{TRUE}}

\item{ServicePoint}{\code{(logical)} Does the CoC use ServicePoint HMIS? \strong{Default} \code{FALSE}}
}
\value{

}
\description{
Setup Rm_data options
}
\examples{
\dontrun{
setup_Rm_data(Clarity = TRUE, ServicePoint = FALSE)
}
}
