% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{setup_RmData}
\alias{setup_RmData}
\title{Setup RmData Options}
\usage{
setup_RmData(
  Clarity = TRUE,
  ServicePoint = FALSE,
  Clarity_URL = "https://cohhio.clarityhs.com",
  clients_to_filter = NULL
)
}
\arguments{
\item{Clarity}{\code{(logical)} Indicates whether the CoC uses Clarity HMIS. \strong{Default} is \code{TRUE}.}

\item{ServicePoint}{\code{(logical)} Indicates whether the CoC uses ServicePoint HMIS. \strong{Default} is \code{FALSE}.}

\item{Clarity_URL}{\code{(character)} The URL for the Clarity HMIS. \strong{Default} is \code{"https://cohhio.clarityhs.com"}.}

\item{clients_to_filter}{\code{(list or NULL)} A list of clients to filter. If \code{NULL}, no filtering is applied.}
}
\value{
This function creates or modifies the \code{.Rprofile} file in the project directory to include HMIS options.
}
\description{
Configure and/or create the \code{.Rprofile} file to set options for the RmData package.
}
\examples{
\dontrun{
setup_RmData(Clarity = TRUE, ServicePoint = FALSE)
}
}
