% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\docType{data}
\name{sp2cl_translations}
\alias{sp2cl_translations}
\title{Service Point / Clarity Field representation translations (numeric/character)}
\format{
An object of class \code{list} of length 5.
}
\usage{
sp2cl_translations
}
\value{
\code{(numeric/character)} Whichever class is opposite the input vector
}
\description{
Service Point / Clarity Field representation translations (numeric/character)
}
\keyword{datasets}
