% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{stop_with_instructions}
\alias{stop_with_instructions}
\title{Stop daily update with an informative error}
\usage{
stop_with_instructions(..., error = FALSE)
}
\arguments{
\item{...}{Error Messages to print to the console}
}
\description{
Throws an error in 00_daily_update.R with additional details
}
