% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{unzip_export}
\alias{unzip_export}
\title{Unzip the HUD Export zip file from the download folder}
\usage{
unzip_export(download_folder = "~/../Downloads/", dir = "data")
}
\arguments{
\item{download_folder}{\code{(character)} path to folder where HUD Exports are downloaded}

\item{dir}{\code{(character)} path of destination directory}
}
\description{
Unzip the HUD Export zip file from the download folder
}
