% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Bed_Unit_Utilization.R
\name{bed_unit_utilization}
\alias{bed_unit_utilization}
\title{Bed Utilizations}
\usage{
bed_unit_utilization(
  Project,
  clarity_api = get_clarity_api(e = rlang::caller_env()),
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{clarity_api}{\code{(clarity_api)} R6 Object containing the connection to the Clarity Looker instance.}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\description{
Bed Utilizations
}
