% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Bed_Unit_Utilization.R
\name{bed_unit_utilization}
\alias{bed_unit_utilization}
\title{Bed Utilizations}
\usage{
bed_unit_utilization(
  clarity_api = get_clarity_api(e = rlang::caller_env()),
  app_env = get_app_env(e = rlang::caller_env())
)
}
\description{
Bed Utilizations
}
