% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{Client_add_UniqueID}
\alias{Client_add_UniqueID}
\title{Add the UniqueID to the Client export}
\usage{
Client_add_UniqueID(
  Client,
  Client_UniqueIDs,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{Client}{\code{(data.frame)} From the HUD CSV Export}

\item{app_env}{}

\item{Client_extras}{\code{(data.frame)} A custom look linking PersonalID & UniqueID}
}
\value{
\code{(data.frame)} with UniqueID column
}
\description{
Add the UniqueID to the Client export
}
