% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Client_redact}
\alias{Client_redact}
\title{Redact PII from Client HUD Export}
\usage{
Client_redact(Client)
}
\arguments{
\item{Client}{\code{(tibble)} Client HUD Export}
}
\value{
\code{(tibble)} Redacted Client HUD Export
}
\description{
This redacts all PII (except DOB) from Client HUD Export
}
\seealso{
Other Client functions: 
\code{\link{Client_filter}()}
}
\concept{Client functions}
