% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update.R
\name{daily_update}
\alias{daily_update}
\title{Update the Rminor apps}
\usage{
daily_update(
  steps = c(`Update Data` = "update", `Run RmData functions` = "funs", `Send
    Dependencies to Dropbox` = "send"),
  funs = rlang::set_names(c("dates", "load_export", "cohorts", "client_counts",
    "qpr_ees", "qpr_spdats", "covid19", "covid19_plots", "vets", "vet_active",
    "prioritization", "bed_unit_utilization", "data_quality", "data_quality_summary")),
  export_zip = TRUE,
  remote = FALSE,
  backup = FALSE,
  app_env,
  clarity_api,
  session,
  e = rlang::caller_env()
)
}
\arguments{
\item{steps}{\code{(character)} vector of the steps to run. Any of:
\itemize{
\item{\code{ update }}{  Update Data }
\item{\code{ funs }}{  Run RmData functions }
\item{\code{ send }}{  Send Dependencies to Dropbox }
}}

\item{funs}{\code{(character)} vector of functions to run. Any of:
\itemize{
\item{\code{ dates }}
\item{\code{ load_export }}
\item{\code{ cohorts }}
\item{\code{ client_counts }}
\item{\code{ qpr_ees }}
\item{\code{ qpr_spdats }}
\item{\code{ covid19 }}
\item{\code{ covid19_plots }}
\item{\code{ vets }}
\item{\code{ vet_active }}
\item{\code{ prioritization }}
\item{\code{ bed_unit_utilization }}
\item{\code{ data_quality }}
\item{\code{ data_quality_summary }}
}}

\item{export_zip}{\code{(logical)} Did the HUD Export CSV's come from an uploaded zip file?}

\item{remote}{\code{(logical)} Whether to send dependencies to the remote storage.}

\item{backup}{\code{(logical)} Whether to back up all raw dependencies in the dependencies environment.}

\item{session}{\code{(shiny session object)}}

\item{e}{\code{(calling environment)}}
}
\value{
Updates the Rminor apps and provides status messages along the way
}
\description{
Update the Rminor apps
}
