% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_summary.R
\name{data_quality_summary}
\alias{data_quality_summary}
\title{Render Data Quality plots & tables for RminorElevated dq_system_wide}
\usage{
data_quality_summary(
  served_in_date_range,
  Referrals,
  Project,
  dq_past_year,
  dq_eligibility_detail,
  dq_overlaps,
  data_APs,
  rm_dates,
  vars,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{Referrals}{\code{(data.frame)} From the Clarity Looker API Extras}

\item{Project}{\code{(data.frame)} From the HUD CSV Export}

\item{dq_past_year}{\code{(data.frame)} See \code{data_quality}}

\item{dq_eligibility_detail}{\code{(data.frame)} See \code{data_quality}}

\item{data_APs}{\code{(data.frame)} See \code{data_APs}}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}
}
\description{
Render Data Quality plots & tables for RminorElevated dq_system_wide
}
