% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_imports.R
\name{data_quality_tables}
\alias{data_quality_tables}
\title{Data Quality Tables Roxygen Template}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{IncomeBenefits}{\code{(data.frame)} From the HUD CSV Export}

\item{Inventory}{\code{(data.frame)} From the HUD CSV Export}

\item{HealthAndDV}{\code{(data.frame)} From the HUD CSV Export}

\item{Disabilities}{\code{(data.frame)} From the HUD CSV Export}

\item{Project}{\code{(data.frame)} From the HUD CSV Export}

\item{Enrollment}{\code{(data.frame)} From the HUD CSV Export}

\item{EnrollmentCoC}{\code{(data.frame)} From the HUD CSV Export}

\item{Exit}{\code{(data.frame)} From the HUD CSV Export}

\item{Client}{\code{(data.frame)} From the HUD CSV Export}

\item{Services}{\code{(data.frame)} From the HUD CSV Export}

\item{Services_extras}{\code{(data.frame)} From the Clarity Looker API Extras}

\item{Referrals}{\code{(data.frame)} From the Clarity Looker API Extras}

\item{mahoning_projects}{\code{(named numeric)} Vector of ProjectIDs associated with projects operating in Mahoning. Created in \code{load_export}.}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Data Quality Tables Roxygen Template
}
