% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_dose_date_error}
\alias{dq_dose_date_error}
\title{Find Dose Date Errors}
\usage{
dq_dose_date_error(
  served_in_date_range,
  vars,
  doses,
  guidance = NULL,
  rm_dates = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Dose Date Errors
}
\seealso{
Other DQ: Vaccines: 
\code{\link{dq_missing_vaccine_current}()},
\code{\link{dq_missing_vaccine_exited}()}
}
\concept{DQ: Vaccines}
