% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_missing_vaccine_current}
\alias{dq_missing_vaccine_current}
\title{Find Missing vaccine data for enrolled clients}
\usage{
dq_missing_vaccine_current(
  served_in_date_range,
  vars,
  dose_counts,
  doses = NULL,
  mahoning_projects = NULL,
  rm_dates = NULL,
  guidance = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{dose_counts}{\code{(data.frame)} Count of doses per client}

\item{doses}{\code{(data.frame)} See the \verb{[["HUD Extras"]]$Client_Doses_extra} method in the instantiated clarity_api object}

\item{mahoning_projects}{\code{(named numeric)} Vector of ProjectIDs associated with projects operating in Mahoning. Created in \code{load_export}.}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Find Missing vaccine data for enrolled clients
}
\seealso{
Other DQ: Vaccines: 
\code{\link{dq_dose_date_error}()},
\code{\link{dq_missing_vaccine_exited}()}
}
\concept{DQ: Vaccines}
