% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_missing_vaccine_exited}
\alias{dq_missing_vaccine_exited}
\title{Data quality report on Race data}
\usage{
dq_missing_vaccine_exited(
  served_in_date_range,
  doses,
  mahoning_projects,
  covid19,
  guidance,
  vars,
  app_env
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{guidance}{\code{(list)}}

\item{vars}{\code{(list)}}

\item{app_env}{\code{(data.frame)} Instead of providing all arguments after the first,\code{app_env} with all arguments saved internally can be provided.}
}
\value{
\code{(data.frame)} with \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Data quality report on Race data
}
