% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{dq_name}
\alias{dq_name}
\title{Data Quality report on Missing First Names}
\usage{
dq_name(
  served_in_date_range,
  guidance = NULL,
  vars = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{served_in_date_range}{\code{(data.frame)} See \code{served_in_date_range}}

\item{guidance}{\code{(list)} See \code{guidance}}

\item{vars}{\code{(named list)}
\itemize{
\item{\code{prep}}{ Column names for Prep}
\item{\code{we_want}}{ Column names for output}
}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} \code{vars$we_want} and \code{Issue} (Issue Name), \code{Type} (Error or Warning), and \code{Guidance} (How to correct the issue)
}
\description{
Data Quality report on Missing First Names
}
\seealso{
Other dq_missing_udes: 
\code{\link{dq_dob}()},
\code{\link{dq_ethnicity}()},
\code{\link{dq_gender}()},
\code{\link{dq_race}()},
\code{\link{dq_ssn}()},
\code{\link{dq_veteran}()}
}
\concept{dq_missing_udes}
