% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Enrollment_add_Household}
\alias{Enrollment_add_Household}
\title{Add Household Information to Enrollment from Project}
\usage{
Enrollment_add_Household(
  Enrollment,
  Project,
  rm_dates,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{Enrollment}{with Exit data. See \code{Enrollment_add_Exit}}

\item{Project}{\code{(data.frame)} From the HUD CSV Export}

\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)} of Enrollment with Household Columns \code{MoveInDateAdjust} appended
}
\description{
Add Household Information to Enrollment from Project
}
