% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{Funder_VA_ProjectID}
\alias{Funder_VA_ProjectID}
\title{Funder_VA}
\usage{
Funder_VA_ProjectID(Funder)
}
\arguments{
\item{Funder}{\code{(data.frame)} From the HUD CSV Export}
}
\value{
\code{(data.frame)} with Project ID
}
\description{
This filters for VA Funded Project IDs
}
