% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{hmis_option}
\alias{hmis_option}
\title{Retrieve an HMIS option set via \code{setup_RmData}}
\usage{
hmis_option(opt, default = FALSE)
}
\arguments{
\item{opt}{\code{(character)} Option name to retrieve}
}
\description{
Retrieve an HMIS option set via \code{setup_RmData}
}
