% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{is_clarity_api}
\alias{is_clarity_api}
\title{Is object of class \code{clarity_api}}
\usage{
is_clarity_api(x)
}
\arguments{
\item{x}{\code{(any)} The object to check.}
}
\value{
\code{(logical)} Returns \code{TRUE} if \code{x} is of class \code{"clarity_api"}, otherwise \code{FALSE}.
}
\description{
Check if the provided object is of class \code{clarity_api}.
}
