% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export.R
\name{load_client}
\alias{load_client}
\title{Load Client data & extras pre-requisites for \code{load_export}}
\usage{
load_client(
  clarity_api = get_clarity_api(e = rlang::caller_env()),
  app_env = get_app_env(e = rlang::caller_env())
)
}
\value{
\code{app_env} with data dependencies
}
\description{
Load Client data & extras pre-requisites for \code{load_export}
}
