% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export.R
\name{load_export}
\alias{load_export}
\title{Load the HUD Export and join extras.}
\usage{
load_export(
  clarity_api = get_clarity_api(e = rlang::caller_env()),
  app_env = get_app_env(e = rlang::caller_env()),
  error = FALSE
)
}
\arguments{
\item{clarity_api}{\code{(clarity_api)} R6 Object containing the connection to the Clarity Looker instance.}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}

\item{error}{\code{(logical)} whether to error or send a message via pushbullet when data checks fail}
}
\value{
\code{app_env} with data dependencies
}
\description{
Load the HUD Export and join extras.
}
