% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_export_utils.R
\name{load_program_lookup}
\alias{load_program_lookup}
\title{Fetch Program Lookup table}
\usage{
load_program_lookup(program_lookup)
}
\arguments{
\item{program_lookup}{Custom program data frame}
}
\value{
\code{(tbl)}
}
\description{
Fetch Program Lookup table
}
