% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export.R
\name{load_public}
\alias{load_public}
\title{Load public data pre-requisites for \code{load_export}}
\usage{
load_public(app_env = get_app_env(e = rlang::caller_env()))
}
\value{
\code{app_env} with data dependencies
}
\description{
Load public data pre-requisites for \code{load_export}
}
