% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update.R
\name{pb_start}
\alias{pb_start}
\title{Start a \link[cli]{cli_progress_bar} or \link[shiny]{Progress} based progress bar}
\usage{
pb_start(
  name = "Daily Update",
  status = NULL,
  type = c("iterator", "tasks", "download", "custom")[1],
  total = NA,
  is_shiny = FALSE
)
}
\arguments{
\item{name}{This is typically used as a label, and should be short,
at most 20 characters.}

\item{status}{New status string of the progress bar, if not \code{NULL}.}

\item{type}{Type of the progress bar. It is used to select a default
display if \code{format} is not specified. Currently supported types:
\itemize{
\item \code{iterator}: e.g. a for loop or a mapping function,
\item \code{tasks}: a (typically small) number of tasks,
\item \code{download}: download of one file,
\item \code{custom}: custom type, \code{format} must not be \code{NULL} for this type.
}}

\item{total}{\code{(numeric)} \code{max} for \link[shiny]{Progress} & \code{total} for \link[cli]{cli_progress_bar}}

\item{is_shiny}{\code{(logical)} Whether running in a shiny environment}
}
\description{
Start a \link[cli]{cli_progress_bar} or \link[shiny]{Progress} based progress bar
}
