% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update_utils.R
\name{pb_update.character}
\alias{pb_update.character}
\title{S3 character Method pb_update}
\usage{
\method{pb_update}{character}(
  pbar,
  message = NULL,
  inc = NULL,
  set = NULL,
  ...,
  e = rlang::caller_env()
)
}
\description{
S3 character Method pb_update
}
