% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update.R
\name{pb_update.Progress}
\alias{pb_update.Progress}
\title{S3 Progress Method pb_update}
\usage{
\method{pb_update}{Progress}(
  pbar,
  message = NULL,
  inc = NULL,
  set = NULL,
  ...,
  e = rlang::caller_env()
)
}
\description{
S3 Progress Method pb_update
}
