% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{pe_add_GrantType}
\alias{pe_add_GrantType}
\title{Add GrantType column to provider_extras}
\usage{
pe_add_GrantType(provider_extras)
}
\value{

}
\description{
GrantType indicates if the program is funded by one of HOPWA, PATH, SSVF, or RHY
}
