% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Pe_add_regions}
\alias{Pe_add_regions}
\title{Add the Corresponding Region for each Project by way of Geocode matching}
\usage{
Pe_add_regions(provider_extras, dirs)
}
\arguments{
\item{dirs}{}
}
\value{
\code{(data.frame)} provider_extras with Regions column
}
\description{
Add the Corresponding Region for each Project by way of Geocode matching
}
