% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{pe_add_regions}
\alias{pe_add_regions}
\title{Add the Corresponding Region for each Project by way of Geocode matching}
\usage{
pe_add_regions(
  provider_extras,
  Regions = clarity.looker::hud_load("Regions", dirs$public),
  dirs
)
}
\arguments{
\item{provider_extras}{\code{(data.frame)} An extra from the from the Program Descriptor model. See \code{?clarity_api} for retrieving a Look with this info.}
}
\value{
\code{(data.frame)} provider_extras with Regions column
}
\description{
Add the Corresponding Region for each Project by way of Geocode matching
}
