% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_Export_utils.R
\name{Pe_create_APs}
\alias{Pe_create_APs}
\title{Add Access Points to Provider_extras}
\usage{
Pe_create_APs(
  provider_extras,
  ProjectCoC,
  dirs,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{provider_extras}{\code{(data.frame)} provider_extras with Regions, see \code{Pe_add_regions}}

\item{dirs}{}
}
\value{
\code{(data.frame)}
}
\description{
Create data.frame of Coordinated Entry Access Points with info about the Counties & Populations Served
}
