% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prioritization.R
\name{prioritization}
\alias{prioritization}
\title{Create the prioritization list}
\usage{
prioritization(
  co_clients_served,
  covid19,
  c19priority,
  Disabilities,
  Enrollment_extra_Exit_HH_CL_AaE,
  HealthAndDV,
  IncomeBenefits,
  Project,
  Referrals,
  Scores,
  lh_project_types,
  ph_project_types,
  clarity_api = get_clarity_api(e = rlang::caller_env()),
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{co_clients_served}{\code{(data.frame)} See \code{cohorts}}

\item{covid19}{\code{(data.frame)} See \code{covid19}}

\item{c19priority}{\code{(data.frame)} See \code{covid19}}

\item{Disabilities}{\code{(data.frame)} See \code{covid19}}

\item{Enrollment_extra_Exit_HH_CL_AaE}{\code{(data.frame)} See \code{load_export}}

\item{IncomeBenefits}{\code{(data.frame)} From the HUD CSV Export}

\item{Project}{\code{(data.frame)} From the HUD CSV Export}

\item{Referrals}{\code{(data.frame)} See \code{load_export}}

\item{Scores}{\code{(data.frame)} See \code{load_export}}

\item{lh_project_types}{\code{(integer)} See \code{cohorts}}

\item{ph_project_types}{\code{(integer)} See \code{cohorts}}

\item{app_env}{\code{(app_env)} R6 Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\description{
Create the prioritization list
}
\seealso{
load_export, covid19, cohorts
}
