% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{projects_current_hmis}
\alias{projects_current_hmis}
\title{Filter for Current HMIS participating projects}
\usage{
projects_current_hmis(
  Project,
  Inventory,
  calc,
  meta_HUDCSV,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{Inventory}{}
}
\value{
\code{(data.frame)}
}
\description{
Filter for Current HMIS participating projects
}
