% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{projects_current_hmis}
\alias{projects_current_hmis}
\title{Filter for Current HMIS participating projects}
\usage{
projects_current_hmis(
  Project,
  Inventory,
  rm_dates,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{rm_dates}{\code{(named list)} with all dates specified or calculated in \code{dates}}

\item{app_env}{\code{(app_env)} Object containing dependencies. If all arguments to this function are saved in the \code{app_env}, then they will be called from there and arguments do not need to be specified.}
}
\value{
\code{(data.frame)}
}
\description{
Filter for Current HMIS participating projects
}
