% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataQuality_utils.R
\name{served_in_date_range}
\alias{served_in_date_range}
\title{Create the data.frame of Clients to Check \code{served_in_date_range}}
\usage{
served_in_date_range(
  projects_current_hmis,
  Enrollment_extra_Exit_HH_CL_AaE = NULL,
  Client = NULL,
  Project = NULL,
  Inventory = NULL,
  HealthAndDV = NULL,
  calc = NULL,
  meta_HUDCSV = NULL,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{projects_current_hmis}{\code{(data.frame)} of Providers to check. See \code{projects_current_hmis}}

\item{Enrollment_extra_Exit_HH_CL_AaE}{\code{(data.frame)} Enrollment with all additions from \code{load_export}}

\item{Client}{\code{(data.frame)} Client with all additions from \code{load_export}}

\item{Project}{\code{(data.frame)} Project with extras including Regions and GrantType see \code{Pe_add_regions} & \code{Pe_add_GrantType}.}

\item{Inventory}{\code{(data.frame)} Inventory}

\item{calc}{\code{(list)} of dates from \code{dates}}

\item{meta_HUDCSV}{\code{(list)} of dates from \code{dates}}

\item{app_env}{\code{(data.frame)} Instead of providing all arguments with NULL defaults,\code{app_env} with all arguments saved internally can be provided.}
}
\value{
\code{(data.frame)}
}
\description{
Create the data.frame of Clients to Check \code{served_in_date_range}
}
