% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_R6.R
\name{setup_RmData}
\alias{setup_RmData}
\title{Setup RmData options}
\usage{
setup_RmData(
  Clarity = TRUE,
  ServicePoint = FALSE,
  Clarity_URL = "https://cohhio.clarityhs.com",
  clients_to_filter = NULL
)
}
\arguments{
\item{Clarity}{\code{(logical)} Does the CoC use Clarity HMIS? \strong{Default} \code{TRUE}}

\item{ServicePoint}{\code{(logical)} Does the CoC use ServicePoint HMIS? \strong{Default} \code{FALSE}}
}
\value{

}
\description{
Setup RmData options
}
\examples{
\dontrun{
setup_RmData(Clarity = TRUE, ServicePoint = FALSE)
}
}
