% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update_utils.R
\name{update_data}
\alias{update_data}
\title{Update HUD CSV Export & Extras data}
\usage{
update_data(clarity_api = RmData::get_clarity_api(e = rlang::caller_env()))
}
\value{
\code{(none)} retrieves and saves Export CSVs via \code{clarity.looker} if not up to date. See \code{\link[clarity.looker]{clarity_api}$get_export} & \code{$get_folder_looks} for details.
}
\description{
Update HUD CSV Export & Extras data
}
